# frontend/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Transactions","description":"Endpoints for creating, deleting, and retrieving financial transactions with pagination"},{"name":"Accounts","description":"Endpoints for managing personal accounts and their balances"},{"name":"Category Budgets","description":"Endpoints for managing monthly spending limits for specific account categories"},{"name":"Analytics","description":"Endpoints for financial data aggregation, summaries, and spending analytics"}],"paths":{"/api/v1/budgets":{"put":{"tags":["Category Budgets"],"summary":"Update an existing category budget limit","description":"Modifies the limit amount for an already existing account budget category.","operationId":"updateLimit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryBudgetUpdateCommand"}}},"required":true},"responses":{"200":{"description":"Budget limit successfully updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryBudgetResponseDto"}}}},"400":{"description":"Invalid input data or negative limit amount"},"404":{"description":"Budget for the specified account and category combination was not found"}}},"post":{"tags":["Category Budgets"],"summary":"Set a new category budget limit","description":"Defines a fresh spending limit for a specific category on a selected account.","operationId":"setCategoryLimit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryBudgetCreateCommand"}}},"required":true},"responses":{"201":{"description":"Budget limit successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryBudgetResponseDto"}}}},"400":{"description":"Invalid input parameters or validation failed"},"404":{"description":"The specified account does not exist"}}}},"/api/v1/transactions":{"get":{"tags":["Transactions"],"summary":"Get paginated transactions","description":"Retrieves a paginated list of transactions, optionally filtered by date range and category.","operationId":"getTransactions","parameters":[{"name":"criteria","in":"query","required":true,"schema":{"$ref":"#/components/schemas/TransactionFilterCriteria"}},{"name":"pageable","in":"query","description":"Pagination and sorting parameters (e.g., page=0, size=10, sort=date,desc)","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Paginated list of transactions successfully retrieved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTransactionResponseDto"}}}}}},"post":{"tags":["Transactions"],"summary":"Create a new transaction","description":"Registers a new financial transaction (Income or Expense). If it is an Expense, it also evaluates the category budget status.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionCommand"}}},"required":true},"responses":{"201":{"description":"Transaction successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponseDto"}}}},"400":{"description":"Invalid input data or budget/account processing failure"},"404":{"description":"The specified account ID does not exist"}}}},"/api/v1/accounts":{"get":{"tags":["Accounts"],"summary":"Get all accounts","description":"Retrieves a full list of all registered personal accounts with their current balances.","operationId":"getAll","responses":{"200":{"description":"Account list successfully retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponseDto"}}}}}}},"post":{"tags":["Accounts"],"summary":"Create a new account","description":"Creates a new personal account with an initial name and balance.","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountCommand"}}},"required":true},"responses":{"201":{"description":"Account successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponseDto"}}}},"400":{"description":"Invalid input data or validation failed"}}}},"/api/v1/budgets/{accountId}":{"get":{"tags":["Category Budgets"],"summary":"Get all budgets for an account","description":"Retrieves a list of all defined category limits associated with a specific account ID.","operationId":"getBudgets","parameters":[{"name":"accountId","in":"path","description":"ID of the account to fetch budgets for","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"List of budgets successfully retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryBudgetResponseDto"}}}}},"404":{"description":"Account with the given ID does not exist"}}}},"/api/v1/accounts/{id}":{"get":{"tags":["Accounts"],"summary":"Get account details","description":"Retrieves detailed information about a specific account by its unique identifier.","operationId":"getAccount","parameters":[{"name":"id","in":"path","description":"Unique identifier of the account","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Account found and details returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponseDto"}}}},"404":{"description":"Account with the given ID does not exist"}}},"delete":{"tags":["Accounts"],"summary":"Delete an account","description":"Deletes an account from the system. Deletion is only permitted if the account has no associated transactions.","operationId":"deleteAccount","parameters":[{"name":"id","in":"path","description":"Unique identifier of the account to delete","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"204":{"description":"Account successfully deleted"},"404":{"description":"Account with the given ID does not exist"},"409":{"description":"Cannot delete account because it contains related transactions"}}}},"/api/v1/accounts/{accountId}/analytics":{"get":{"tags":["Analytics"],"summary":"Get financial analytics for an account","description":"Calculates total income, total expenses, and groups expenses by category within a specified time range for a given account.","operationId":"getAccountAnalytics","parameters":[{"name":"accountId","in":"path","description":"ID of the account to analyze","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"fromDate","in":"query","description":"Start date-time of the analysis period (ISO format)","required":true,"schema":{"type":"string","format":"date-time"},"example":"2026-06-01T00:00:00"},{"name":"toDate","in":"query","description":"End date-time of the analysis period (ISO format)","required":true,"schema":{"type":"string","format":"date-time"},"example":"2026-06-30T23:59:59"}],"responses":{"200":{"description":"Financial analytics successfully calculated and retrieved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountAnalyticsResponseDto"}}}},"400":{"description":"Invalid date-time format or missing request parameters"},"404":{"description":"Account with the specified ID was not found"}}}},"/api/v1/transactions/{id}":{"delete":{"tags":["Transactions"],"summary":"Delete a transaction","description":"Removes a specific transaction by its unique database ID and automatically reverts its impact on the account balance.","operationId":"delete","parameters":[{"name":"id","in":"path","description":"Unique identifier of the transaction","required":true,"schema":{"type":"integer","format":"int64"},"example":45}],"responses":{"204":{"description":"Transaction successfully deleted"},"404":{"description":"Transaction with the given ID does not exist"}}}},"/api/v1/budgets/{id}":{"delete":{"tags":["Category Budgets"],"summary":"Delete a budget limit","description":"Removes a specific category budget limit from the system completely using its unique database ID.","operationId":"deleteLimit","parameters":[{"name":"id","in":"path","description":"Unique identifier of the budget limit configuration","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"204":{"description":"Budget limit successfully removed"},"404":{"description":"Budget limit with the given ID does not exist"}}}}},"components":{"schemas":{"CategoryBudgetUpdateCommand":{"type":"object","description":"Data required to update an existing category spending limit","properties":{"accountId":{"type":"integer","format":"int64","description":"ID of the account associated with the budget","example":1},"category":{"type":"string","description":"The targeted budget category to modify","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"],"example":"ENTERTAINMENT"},"limitAmount":{"type":"number","description":"The new maximum spending threshold. Must be greater than zero.","example":650.0}},"required":["accountId","category","limitAmount"]},"CategoryBudgetResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountId":{"type":"integer","format":"int64"},"category":{"type":"string","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"]},"limitAmount":{"type":"number"}}},"CreateTransactionCommand":{"type":"object","description":"Data required to create a new income or expense transaction","properties":{"amount":{"type":"number","description":"The monetary value of the transaction. Must be positive.","example":250.5},"transactionType":{"type":"string","description":"The type of the transaction","enum":["INCOME","EXPENSE"],"example":"EXPENSE"},"category":{"type":"string","description":"The financial category associated with this transaction","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"],"example":"FOOD"},"description":{"type":"string","description":"Optional custom description or notes regarding the transaction","example":"Weekly grocery shopping at Lidl","maxLength":255,"minLength":0},"date":{"type":"string","format":"date-time","description":"The date and time when the transaction occurred (ISO format)","example":"2026-06-02T12:00:00"},"accountId":{"type":"integer","format":"int64","description":"The database identifier of the account this transaction belongs to","example":1}},"required":["accountId","amount","category","date","transactionType"]},"TransactionResponseDto":{"type":"object","description":"Detailed response information representing a registered transaction and its current category budget impacts","properties":{"id":{"type":"integer","format":"int64","description":"The unique database identifier of the transaction","example":45},"amount":{"type":"number","description":"The monetary value of the transaction","example":250.5},"type":{"type":"string","description":"The type classification of the transaction","enum":["INCOME","EXPENSE"],"example":"EXPENSE"},"category":{"type":"string","description":"The specific spending or income category","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"],"example":"FOOD"},"description":{"type":"string","description":"The description or note assigned to this entry","example":"Weekly grocery shopping at Lidl"},"date":{"type":"string","format":"date-time","description":"The registered date and time of the transaction","example":"2026-06-02T12:00:00"},"accountId":{"type":"integer","format":"int64","description":"The identifier of the associated account","example":1},"budgetExceeded":{"type":"boolean","description":"Flag indicating if the current month's budget limit for this category has been exceeded after inserting this transaction","example":true},"spentInCategory":{"type":"number","description":"Total accumulated spending amount in this specific category for the current month so far","example":1450.75}}},"CreateCategoryBudgetCreateCommand":{"type":"object","description":"Data required to set a new spending limit configuration for a category","properties":{"accountId":{"type":"integer","format":"int64","description":"ID of the account to assign the budget to","example":1},"category":{"type":"string","description":"The target category for the limit","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"],"example":"FOOD"},"limitAmount":{"type":"number","description":"Maximum allowed spending amount for this category. Must be greater than zero.","example":1200.0}},"required":["accountId","category","limitAmount"]},"CreateAccountCommand":{"type":"object","properties":{"name":{"type":"string","description":"The unique name of the account","example":"my account"}},"required":["name"]},"AccountResponseDto":{"type":"object","description":"Response data containing details of a personal account and its current balance","properties":{"id":{"type":"integer","format":"int64","description":"The unique database identifier of the account","example":1},"name":{"type":"string","description":"The name given to the account","example":"Main Checking Account"},"balance":{"type":"number","description":"The current monetary balance available in the account","example":3500.0}}},"TransactionFilterCriteria":{"type":"object","description":"Criteria parameters used for filtering the paginated transaction list","properties":{"from":{"type":"string","format":"date-time"},"to":{"type":"string","format":"date-time"},"category":{"type":"string","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"]}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageTransactionResponseDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponseDto"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"AccountAnalyticsResponseDto":{"type":"object","description":"Response data containing comprehensive financial analytics and summaries for a specific account","properties":{"accountId":{"type":"integer","format":"int64","description":"The analyzed account identifier","example":1},"period":{"$ref":"#/components/schemas/PeriodDto","description":"The time range period for which the analytics were calculated"},"totalIncome":{"type":"number","description":"Total accumulated income within the requested period","example":5200.0},"totalExpenses":{"type":"number","description":"Total accumulated expenses within the requested period","example":1840.5},"expensesByCategory":{"type":"array","description":"Breakdown of total expenses grouped by each individual category","items":{"$ref":"#/components/schemas/ExpenseByCategoryDto"}}}},"ExpenseByCategoryDto":{"type":"object","description":"Aggregated details of expenses for a specific category, including its percentage share of total spending","properties":{"category":{"type":"string","description":"The financial category of the expense","enum":["HOUSING","FOOD","TRANSPORTATION","ENTERTAINMENT","SALARY","UNCATEGORIZED"],"example":"FOOD"},"amount":{"type":"number","description":"The total monetary amount spent in this category within the period","example":450.75},"percentage":{"type":"number","description":"The percentage share of this category relative to the total expenses (0.00 - 100.00)","example":24.5}}},"PeriodDto":{"type":"object","description":"The precise date and time boundaries used to filter and aggregate the financial data","properties":{"fromDate":{"type":"string","format":"date-time","description":"The starting date-time boundary of the analysis period (inclusive)","example":"2026-06-01T00:00:00"},"toDate":{"type":"string","format":"date-time","description":"The ending date-time boundary of the analysis period (inclusive)","example":"2026-06-30T23:59:59"}}}}}}